+++ clox/src/debug.h
@@ -6,7 +6,7 @@
 void
 disassemble_chunk(Chunk *chunk, const char *name);
 
-size_t
-disassemble_instruction(Chunk *chunk, size_t offset);
+int
+disassemble_instruction(Chunk *chunk, int offset);
 
 #endif
